% Generated by roxygen2 (4.0.1): do not edit by hand
\name{daj_grupowanie}
\alias{daj_grupowanie}
\title{Pobieranie nazw zmiennych z formuł}
\usage{
daj_grupowanie(dane, zmiennaGr)
}
\arguments{
\item{dane}{tablica z danymi.}

\item{zmiennaGr}{wektor nazw kolumn tablicy dane, dla których mam}
}
\value{
Funkcja zwraca tablicę, której wiersze zawierają wszystkie kombinacje poziomów zmiennych grupujacych.
}
\description{
Funkcja pomocnicza, która zwraca nazwy prawostronnych zmiennych formuły. Jeżeli formuła jest obustronna to funkcja zwraca błąd.
}

