% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wyliczanieEWD.R
\name{ewd_es}
\alias{ewd_es}
\title{Wyliczanie EWD (Kalkulator)}
\usage{
ewd_es(x, idSzkoly)
}
\arguments{
\item{x}{model klasy \code{lm}}

\item{idSzkoly}{data frame zawierający kolumnę z identyfikatorami szkół}
}
\value{
data frame zawierający oszacowania dla poszczególnych szkół: EWD, błędu
standardowego EWD, średniego wyniku końcowegoi jego błędu standardowego
}
\description{
Funkcja zwraca oszacowania EWD i ich błędy standardowe wyliczane jako średnia z reszt
regresji MNK oraz średnie wyniki egzaminu na wyjściu i ich błędy standardowe wyliczane
jako błąd standardowy średniej z prostej próby losowej.
}
\details{
Ponieważ obiekt z modelem regresji MNK nie zawiera informacji o przysziale uczniów do
szkół, musi ona zostać podana oddzielnym parametrem (\code{id_szkoly}). Musi to być
data frame, zawierający kolumnę z id szkoły (i najlepiej tylko nią jedną), utworzony
przez usunięcie kolumn z tego samego obiektu, z którego był estymowany model \code{x}.
Łączenie jest przeprowadzane po nazwach wierszy (\code{model.frame(x)})
i (\code{id_szkoly}).
}

