% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ewd_me}
\alias{ewd_me}
\title{Wyliczanie EWD}
\usage{
ewd_me(x)
}
\arguments{
\item{x}{model klasy \code{lmerMod}}
}
\value{
data frame
}
\description{
Funkcja zwraca oszacowania EWD i ich "błędy standardowe" wykorzystując funkcję
\code{\link[EWDogolny]{ranef_ddf}}.
}

