% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ewd_me_ssr}
\alias{ewd_me_ssr}
\title{Wyliczanie EWD}
\usage{
ewd_me_ssr(x)
}
\arguments{
\item{x}{model klasy \code{lmerMod}}
}
\value{
data frame
}
\description{
Funkcja zwraca oszacowania EWD i ich "błędy standardowe" wykorzystując metodę
"ściągniętych średnich reszt".
}
\details{
Funkcja powstała jako awaryjne rozwiązanie w sytuacji, gdy
\code{ranef(., condVar=TRUE)} nie dawało dobrych oszacowań "błędów standardowych"
ze względu na błąd autorów pakietu \code{lme4}.
}

