% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lme_ewd.R
\name{konwertujNalmeEWD}
\alias{konwertujNalmeEWD}
\title{Konwersja obiektu lme4 na lmeEWD}
\usage{
konwertujNalmeEWD(objectLME4)
}
\arguments{
\item{objectLME4}{model klasy \code{lmerMod}}
}
\value{
obiekt klasy 'lmeEWD'.
}
\description{
Funkcja tworzy obiekt klasy 'lmeEWD' na podstawie obiektu 'lmerMod'.
}

