% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koryguj_wskazniki_ewd.R
\name{koryguj_wsk_ewd}
\alias{koryguj_wsk_ewd}
\title{Obliczanie EWD}
\usage{
koryguj_wsk_ewd(modele, dane, daneKorekty, skale = NULL,
  powiazaniaPrzedmiotow = NULL)
}
\arguments{
\item{modele}{lista modeli klasy \code{lm} lub \code{lmerMod}}

\item{dane}{data frame z danymi, na podstawie
których estymowane były modele podane parametrem \code{modele}}

\item{daneKorekty}{data frame z danymi szkół, dla których mają zostać
przeliczone wskaźniki}

\item{skale}{data frame, która zostanie przekazana
jako argument o tej samej nazwie do \link{przygotuj_wsk_ewd}}

\item{powiazaniaPrzedmiotow}{opcjonalnie lista, która zostanie przekazana
jako argument o tej samej nazwie do \link{przygotuj_wsk_ewd};}
}
\value{
lista data frame'ów
}
\description{
Funkcja na podstawie listy modeli, data frame'a z danymi, na podstawie
których zostały one obliczone oraz data frame'a z danymi skorygowanymi,
dla których wartości wskaźników EWD mają zostać przeliczone, przygotowuje
zestawienia z wartościami wskaźników EWD (używając jako konia roboczego
\link{przygotuj_wsk_ewd}).
}
