% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lines.wielomian}
\alias{lines.wielomian}
\title{Przeciążenie funkcji lines dla obiektu wielomian.}
\usage{
\method{lines}{wielomian}(wielomian, colRos = 3, colMal = 4, lwd = 2,
  lty = 1)
}
\arguments{
\item{wielomian}{obielt klasy 'wielomian'.}

\item{colRos}{kolor rosnącej części wykresu wielomianu.}

\item{colMal}{kolor malejacej części wykresu wielomianu.}

\item{lwd}{szerokość linii. Patrz również \code{\link[graphics]{par}}.}

\item{lty}{typ linii. Patrz również \code{\link[graphics]{par}}.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja dodaje od wykresy linię wielomianu.
}

