% Generated by roxygen2 (4.0.1): do not edit by hand
\name{model.map}
\alias{model.map}
\title{Mapowanie zmiennych dla modelu z interakcjami}
\usage{
model.map(model, zmiennaWielomianowa)
}
\arguments{
\item{model}{parametr klasy 'lmerMod' lub 'lm'.}

\item{zmiennaWielomianowa}{nazwa zmiennej, która jest parametrem wielomianu.}
}
\value{
funkcja zwraca macierz logiczną, której nazwy wierszy to zmienne ramki danych, a nazwy kolumn to nazwy współczynników modelu.
}
\description{
Celem funkcji jest stworzenie tablicy, która zawiera mapowanie kolumn ramki danych, na której był budowany model
na nazwy współczynników modelu (i macierzy modelu).
}

