% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostyka_wiele_fun_pom.R
\name{model_map}
\alias{model_map}
\title{Mapowanie zmiennych dla modelu z interakcjami}
\usage{
model_map(model, zmiennaWielomianowa)
}
\arguments{
\item{model}{parametr klasy 'lmerMod' lub 'lm'.}

\item{zmiennaWielomianowa}{nazwa zmiennej, która jest parametrem wielomianu.}
}
\value{
funkcja zwraca macierz logiczną, której nazwy wierszy to zmienne
ramki danych, a nazwy kolumn to nazwy współczynników modelu
}
\description{
Celem funkcji jest stworzenie tablicy, która zawiera mapowanie kolumn ramki
danych, na której był budowany model na nazwy współczynników modelu
(i macierzy modelu).
}

