% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mojeParametry}
\alias{mojeParametry}
\title{Parametry rozkładów zmiennych.}
\usage{
mojeParametry(x, na.rm = TRUE, digits = 2)
}
\arguments{
\item{x}{wektor liczbowy}

\item{na.rm}{wartość logiczna - przekazywana do funkcji wyliczających statystyki}

\item{digits}{liczba całkowita - do ilu miejsc po przecinku zaokrąglać zwracane wyniki}
}
\value{
wektor liczb
}
\description{
Funkcja zwraca wartości zestawu statystyk opisowych: kwartyle, średnia, odch. stand.
}

