% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/panoramaEwd.R
\name{panorama_ewd}
\alias{panorama_ewd}
\title{Diagnostyka wskaźników EWD}
\usage{
panorama_ewd(sr, ewd, egzamin, typ_szkoly, wskaznik, lata, zapiszPng = NULL,
  xlab = wskaznik, paleta = colorRampPalette(c("white", blues9)),
  prElips = NULL)
}
\arguments{
\item{sr}{wektor liczbowy z wartościami średnich wyników egzaminu na wyjściu
poszczególnych szkół}

\item{ewd}{wektor liczbowy z wartościami wskaźników EWD szkół}

\item{egzamin}{ciąg znaków z nazwą egzaminu (w dopełniaczu)}

\item{typ_szkoly}{ciąg znaków z typem szkoły (typowo "gimn.", "LO" lub "T")}

\item{wskaznik}{ciąg znaków z nazwą wskaźnika EWD}

\item{lata}{ciąg znaków opisujący, jakich lat wskaźnik dotyczy}

\item{zapiszPng}{ciąg znaków - ścieżka do katalogu, w którym mają być zapisane pliki
PNG z wykresami lub NULL, jeśli wykresy mają nie być zapisywane}

\item{xlab}{opcjonalnie ciąg znaków - podpis osi X, jeśli ma być inny niż
\code{wskaznik}}

\item{paleta}{funkcja definiująca paletę kolorów do wykorzystania przy rysowaniu
wykresów, typowo wynik wywołania funkcji \code{\link[grDevices]{colorRampPalette}}}

\item{prElips}{wektor liczb z przedziału (0;1), które oznaczają  odsetek szkół, które mają się znaleźć w elipsie.}
}
\value{
funkcja nic nie zwraca.
}
\description{
Funkcja rysuje "panoramę polskiej edukacji", czyli wykres rozrzutu szkół ze względu na
wartości średnich wyników na wyjściu (oś X) i wartości wskaźników EWD (oś Y).
}

