% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_ewd.R
\name{panorama_z_listy_wskaznikow}
\alias{panorama_z_listy_wskaznikow}
\title{Diagnostyka wskaźników EWD}
\usage{
panorama_z_listy_wskaznikow(x, lata, katalogZapis = NULL,
  wybierzSzkoly = NULL, typSzkol = NULL)
}
\arguments{
\item{x}{lista data frame'ów zwracana przez funkcję
\code{\link{przygotuj_wsk_ewd}}}

\item{lata}{ciąg znaków opisujący, jakich lat wskaźnik dotyczy}

\item{katalogZapis}{opcjonalnie ciąg znaków podający nazwę katalogu, do
którego mają zostać zapisane rysunki (jeśli nie zostanie podana, rysunki
nie zostaną zapisane)}

\item{wybierzSzkoly}{opcjonalnie wyrażenie (expression), które zostanie
użyte do zawężenia grupy szkół, których wyniki będą prezentowane; np.
\code{expression(lu_wszyscy >= 10)}}

\item{typSzkol}{opcjonalnie ciąg znaków z nazwą typu szkół (potrzebny tylko
w przypadku matury)}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja rysuje "panoramę polskiej edukacji", czyli wykres rozrzutu szkół ze
względu na wartości średnich wyników "na wejściu"/"na wyjściu" (oś X)
i wartości wskaźników EWD (oś Y).
}
