% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porownaj_wielomiany.R
\name{porownaj_wielomiany}
\alias{porownaj_wielomiany}
\title{Diagnostyka wielomianow}
\usage{
porownaj_wielomiany(x, zapis = NULL, cyfryLinkTest = 5, cyfryR2 = 4)
}
\arguments{
\item{x}{lista obiektów zwracanych przez funkcję
\code{\link{diagnostyka_wielomianow}}}

\item{zapis}{opcjonalnie nazwa pliku, do którego ma zostać zapisane
zestawienie}

\item{cyfryLinkTest}{liczba cyfr po przecinku, do której zostaną zaokrąglone
istotności "linktestu"}

\item{cyfryR2}{liczba cyfr po przecinku, do której zostaną zaokrąglone
wartości R2 (i BIC)}
}
\value{
Funkcja zwraca listę z zestawieniami: 1) wyników "linktestu",
2) sprawdzenia monotoniczności, 3) parametrami BIC i R2 modeli
}
\description{
Funkcja przygotowuje zestawienie porównujące modele z wielomianami różnych
stopni.
}

