% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lme_ewd.R
\name{predict.lmeEWD}
\alias{predict.lmeEWD}
\title{Przewidywanie dla obiektu 'lmeEWD'}
\usage{
\method{predict}{lmeEWD}(object, ...)
}
\arguments{
\item{object}{model klasy \code{lmeEWD}}

\item{...}{parametry: 'newdata' - ramka danych, dla których należy wyliczyć przewidywania oraz
'zLosowymi' -  wartość boolowska określająca, czy uwzględniać w przewidywaniach efekt losowy.}
}
\value{
wektor liczb.
}
\description{
Funkcja tworzy obiekt klasy 'lmeEWD' na podstawie obiektu 'lmerMod'.
}

