% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przeksztalc_skale_zmiennych.R
\name{przeksztalc_dane_normami}
\alias{przeksztalc_dane_normami}
\title{Wyliczanie EWD}
\usage{
przeksztalc_dane_normami(dane, normy)
}
\arguments{
\item{dane}{data frame z danymi, typowo wczytany do R wynik działania
funkcji \code{\link[EWDdane]{przygotuj_dane_do_ewd}}}

\item{normy}{lista data frame'ów z normami, typowo wynik działania funkcji
\code{\link{przeksztalc_skale_norm}}}
}
\value{
data frame
}
\description{
Funkcja przekształca wyniki egzaminów wyrażone na skalach raschowych
na podstawie podanych norm. Typowo na potrzeby jednorocznych modeli
maturalnych.
}
\details{
W danych muszą znajdować się zmienne opisujące przydział obserwacji do grup
(o nazwach postaci \code{grupa_s}, gdzie 's' to kod rodzaju egzaminu:
's' - sprawdzian, 'g' - egz. gimn., 'm' - matura) i surowe wyniki skal
(o nazwach postaci \code{konstrukt_suma}, gdzie 'konstrukt' to nazwa
konstruktu estymowanego w ramach skali - modelem raschowym). Jeśli w danych
znajdują się już zmienne opisujące wyniki wyskalowane (o nazwach postaci
\code{konstrukt_irt}), zostaną one zastąpione.
}
