% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przeksztalc_skale_zmiennych.R
\name{przeksztalc_skale_norm}
\alias{przeksztalc_skale_norm}
\title{Wyliczanie EWD}
\usage{
przeksztalc_skale_norm(normy, parametry, na10015 = TRUE)
}
\arguments{
\item{normy}{lista data frame'ów z normami, typowo atrybut \code{normy}
wczytanego do R wynik działania funkcji
\code{\link[EWDdane]{przygotuj_dane_do_ewd}}}

\item{parametry}{data frame z parametrami (średnia, odchylenie standardowe)
zmiennych do przekształcenia, typowo wynik działania funkcji
\code{\link{parametry_rozkladu_rasch}}}

\item{na10015}{wartość logiczna - czy po przekształceniu zgodnie z podanymi
parametrami przemnożyć wartości wszystkich norm przez 15 i dodać 100?
(tj. przekształcić na skalę 100;15, zakładając, że po ew. przekształceniach
są wystandaryzowane do średniej 0 i odch. std. 1)}
}
\value{
lista o strukturze zgodnej ze strukturą argumentu \code{normy}
}
\description{
Funkcja przekształca skale norm zgodnie z podanymi parametrami. Typowo na
potrzeby jednorocznych modeli maturalnych, gdzie zachodzi ten problem, że
wyniki egzaminu gimnazjalnego nie są wystandaryzowane względem poszczególnych
typów szkół ponadgimnazjalnych.
}

