% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/przeksztalc_skale_zmiennych.R
\name{przeksztalc_skale_zmiennych}
\alias{przeksztalc_skale_zmiennych}
\title{Wyliczanie EWD}
\usage{
przeksztalc_skale_zmiennych(dane, czesciEgzWy, sufiksWyniki)
}
\arguments{
\item{dane}{data frame z danymi, typowo wczytany do R wynik działania
funkcji \code{\link[EWDdane]{przygotuj_dane_do_ewd}}}

\item{czesciEgzWy}{wektor ciągów znaków - kody konstruktów, które będą
w modelach EWD zmiennymi zależnymi}

\item{sufiksWyniki}{ciąg znaków: "irt", "norm" lub "suma"; określa, które
wyniki będą używane do wyliczania wskaźników i mają zostać przekształcone}
}
\value{
data frame
}
\description{
Funkcja na podstawie listy modeli, data frame'a z danymi, na podstawie
których zostały one wyliczone oraz ew. data frame'a o takiej samej
strukturze zawierającego dane nie użyte wcześniej do estymacji modeli,
przygotowuje zestawienia z wartościami wskaźników EWD.
}

