% Generated by roxygen2 (4.0.1): do not edit by hand
\name{przewidywanie_lmer}
\alias{przewidywanie_lmer}
\title{Przeciążenie funkcji lines dla obiektu wielomian.}
\usage{
przewidywanie_lmer(model, wielomian, grupa, zmienna)
}
\arguments{
\item{model}{obiekt klasy 'lmerMod' lub 'lm'.}

\item{wielomian}{obiekt klasy 'wielomian'.}

\item{grupa}{wektor wartości logiczych określający czy obserwacja należy do grupy.}

\item{zmienna}{ciąg znaków określający nazwę zmiennej wielomianu.}
}
\value{
Funkcja zwraca ramkę danych, która zawiera informacje dla obserwacji o:
grupowaniu, zmiennej zależnej, zmiennej niezależnej, efektach losowych, resztach oraz efetach związanych z wielomianem.
}
\description{
Funkcja dodaje od wykresy linię wielomianu.
}

