% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obliczanie_ewd.R
\name{przygotuj_wsk_ewd}
\alias{przygotuj_wsk_ewd}
\title{Obliczanie EWD}
\usage{
przygotuj_wsk_ewd(modele, dane, danePominiete = NULL, skale = NULL,
  powiazaniaPrzedmiotow = NULL)
}
\arguments{
\item{modele}{lista modeli klasy \code{lm} lub \code{lmerMod}}

\item{dane}{data frame z dodatkowymi danymi, w szczególności zawierający
kolumnę \code{lu_wszyscy}, przypisanymi tym samym obserwacjom, na podstawie
których estymowane były modele podane parametrem \code{modele}}

\item{danePominiete}{opcjonalnie data frame o takiej samej strukturze, jak
przekazany argumentem \code{dane}, zawierający dane szkół, które nie zostały
wykorzystane na etapie estymacji modelu, ale chcemy teraz obliczyć dla nich
wartości wskaźników}

\item{skale}{data frame zawierający informacje o skalowaniach (i skalach),
z których pochodzą zmienne z oszacowaniami umiejętności, zawarte w danych;
typowo atrybut \code{skale} obiektu klasy \code{daneWyskalowane}
(zwracanego przez funkcję \code{\link[EWDdane]{przygotuj_dane_do_ewd}})}

\item{powiazaniaPrzedmiotow}{opcjonalnie lista, której elementy są wektorami
tekstowymi zawierającymi sufiksy nazw przedmiotów, a nazwy tych elementów
odopwiadają sufiksom modeli przekazanych argumentem \code{modele}; pozwala
określić, jakie informacje o liczbie ucznióW powinny zostać zapisane
w zwracanych wynikach; jeśli nie zostanie podany, funkcja spróbuje użyć
domyślnego mapowania, zapisanego w jej kodzie;}
}
\value{
lista data frame'ów
}
\description{
Funkcja na podstawie listy modeli, data frame'a z danymi, na podstawie
których zostały one obliczone oraz ew. data frame'a o takiej samej
strukturze zawierającego dane nie użyte wcześniej do estymacji modeli,
przygotowuje zestawienia z wartościami wskaźników EWD.
}

