% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reszta_grupa}
\alias{reszta_grupa}
\title{Obliczanie reszt dla grupy}
\usage{
reszta_grupa(wielomian, matrix, frame, ranefModel, fixefModel)
}
\arguments{
\item{wielomian}{obiekt klasy 'wielomian'.}

\item{matrix}{macierz modelu.}

\item{frame}{wejściowa ramka danych do modelu.}

\item{ranefModel}{lista efektóe losowych modelu.}

\item{fixefModel}{lista efektów stałych}
}
\value{
Funkcja zwraca ramkę danych, która zawiera informacje dla obserwacji o:
grupowaniu, zmiennej niezależnej, efektach losowych, resztach oraz efetach związanych z wielomianem.
}
\description{
Funkcja oblicza reszt związane z efektami stałymi.
}

