% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sklej_normy.R
\name{sklej_normy}
\alias{sklej_normy}
\title{Przeksztalcanie normalizacji ekwikwantylowej}
\usage{
sklej_normy(wynikiZnorm, wynikiSurowe, do, grupy = NULL, sr = 100,
  os = 15, ile = do)
}
\arguments{
\item{wynikiZnorm}{wektor wartości na skali znormalizowanej}

\item{wynikiSurowe}{wektor wartości surowych, odpowiadających wartościom
znormalizowanym podanym w \code{wynikiZnorm}}

\item{do}{wartość liczbowa - wartość surowa, do jakiej należy kontynuować
skracanie (ale nie zostanie przeprowadzonych więcej złączeń, niż wartość
argumentu \code{ile})}

\item{grupy}{opcjonalnie zmienna definiująca podział na grupy, w ramach
których oddzielnie ma być przeprowadzone łączenie}

\item{sr}{liczba - średnia znormalizowanej skali}

\item{os}{liczba - odchylenie standardowe znormalizowanej skali}

\item{ile}{wartość liczbowa - maksymalna liczba złączeń, które mogą zostać
wykonane}
}
\value{
Wektor liczbowy z przeliczonymi wartościami argumentu \code{wynikiZnorm}.
}
\description{
Funkcja służy do sklejania ze sobą najniższych wartości skali znormalizowanej
ekwikwantylowo, przeliczając jednocześnie wartości skali.
}
\details{
Tu w przyszłości uzupełnić. Trzeba dodać do funkcji jakąś weryfikację
poprawności argumentów.
}
\seealso{
\code{\link[ZPD]{normalizuj}}
}

