% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sr_wy}
\alias{sr_wy}
\title{Wyliczanie EWD}
\usage{
sr_wy(model, ewd)
}
\arguments{
\item{model}{model klasy \code{lmerMod}}

\item{ewd}{data frame będący wynikiem działania funkcji \code{\link{ewd_me}} lub \code{\link{ewd_me_ssr}}}
}
\value{
data frame
}
\description{
Funkcja zwraca oszacowania średnich wyników na wyjściu i ich "błędy standardowe"
oraz korelacje z EWD wykorzystując metodę "z jednego modelu".
}

