% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabelki_dane_do_modeli_ewd.R
\name{tabelki_dane_do_modeli_ewd}
\alias{tabelki_dane_do_modeli_ewd}
\title{Liczba zdających egzaminy.}
\usage{
tabelki_dane_do_modeli_ewd(x, czesciEgzWy, czesciEgzWe)
}
\arguments{
\item{x}{data frame z danymi potrzebnymi do przygotowania zestawienia}

\item{czesciEgzWy}{wektor ciągów znaków - kody części egzaminu "wyjściu"
(np. "gh_h")}

\item{czesciEgzWe}{- wektor ciągów znaków - kody części egzaminu "wejściu"
(np. "s")}
}
\value{
lista data frame'ów
}
\description{
Funkcja przygotowuje zestaw tabelek z zestawieniem liczby podchodzących do
egzaminów i statystykami opisowymi wyników.
}
\details{
.
}

