% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_ewd.R
\name{wielkoscWarstwic}
\alias{wielkoscWarstwic}
\title{Wyliczanie EWD}
\usage{
wielkoscWarstwic(wyniki, ewd, liczbaUczniow, pr = c(0.5, 0.9))
}
\arguments{
\item{wyniki}{wektor z wartościami średnich wyników na wyjściu poszczególnych
szkół}

\item{ewd}{wektor z wartościami wskaźników EWD szkół}

\item{liczbaUczniow}{wektor z liczbą uczniów z poszczególnych szkołach}

\item{pr}{wektor prawdopodobieństw, dla których mają być wyznaczone
odpowiadające im prawdopodobieństwa empiryczne}
}
\value{
wektor liczbowy
}
\description{
Funkcja zwraca "empiryczne prawdopodobieństwa" pozwalające tak dobrać
wielkość warstwic, aby znajdował się w ich obrębie założony odsetek uczniów
(choć z założenia, że rozkład wyników na wyjściu i EWD jest dwuwymiarowym
rozkładem normalnym wynikałaby nieco inna wielkość elips).
}
