% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wielomian_grupa}
\alias{wielomian_grupa}
\title{Wyliczenia dla wielomianu.}
\usage{
wielomian_grupa(matrix, fixefModel, zmienna, poziomyGrup)
}
\arguments{
\item{matrix}{macierz danych badanego modelu.}

\item{fixefModel}{wektor współczynników efektów stałych modelu.}

\item{zmienna}{ciąg znaków określający zmienną wielomianu.}

\item{poziomyGrup}{wiersz ramki danych z poziomami, dla której wykonywane są obliczenia.
Nazwy kolumn opisują zmienne.}
}
\value{
Funkcja zwraca ramkę danych z wartościami zmiennej, obliczeniami reszt oraz wartościami przewidywań na podstawie wielomianu.
Atrybuty ramki danych zawierają informacje o wartościach współczynników przy kolejnych potęgach, informację czy funkcja jest roasnąca oraz
nazwy zmiennych z ramki danych, które brały udział w obliczeniach współczynnika wielomianu.
}
\description{
Funkcja wykonująca wyliczenia dla wielomianowej części modelu w zadanej grupie obserwacji.
}

