% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wielomian_lmer}
\alias{wielomian_lmer}
\title{Wyliczenia dla wielomianu.}
\usage{
wielomian_lmer(model, zmienna, zmiennaGr, grupa, grupowanie, indeksGrupowania,
  nsiatka = 0, czyTylkoGrupa = FALSE)
}
\arguments{
\item{model}{obiekt klasy 'lm' lub 'lmerMod'.}

\item{zmienna}{ciąg znaków określający zmienną wielomianu.}

\item{zmiennaGr}{wektor ciągów znakowych okreslających zmienne modelu, na podstawie których określone są grupy, w których wykonywane są obliczenia oraz dla których są sporządzane wykresy.}

\item{grupa}{ciąg zmiennych logicznych określający, która obserwacja znajduje się w rozważanej grupie.}

\item{grupowanie}{tablica, która jest kombinacją poziomów zmiennych wskazanych przez zmiennaGr.}

\item{indeksGrupowania}{indeks tablicy podawanej w parametrze 'grupowanie'. Indeks wskazuje bierzącą grupę.}

\item{nsiatka}{liczba punktów, w której ma zostać obliczony wielomian. Domyślna wartość 0 oznacza, że wielomian będzie obliczony dla każdej występującej wartości.}

\item{czyTylkoGrupa}{wartośc logiczna. Jeżeli wynosi TRUE to wartości oblicznanego wielomianu są między minimum i maksimum dla grupy.
Jeżeli wartość wynosi FALSE to minimum i maksimum jest liczone dla całej populacji.}
}
\value{
Funkcja zwraca ramkę danych z wartościami zmiennej, obliczeniami reszt oraz wartościami przewidywań na podstawie wielomianu.
Atrybuty ramki danych zawierają informacje o wartościach współczynników przy kolejnych potęgach, informację czy funkcja jest rosnąca oraz
nazwy zmiennych z ramki danych, które brały udział w obliczeniach współczynnika wielomianu.
}
\description{
Funkcja wykonująca wyliczenia dla wielomianowej części modelu w zadanej grupie obserwacji.
}

