% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wyciagnij_nazwe_zmiennej}
\alias{wyciagnij_nazwe_zmiennej}
\title{Pobieranie nazw zmiennych z formuł}
\usage{
wyciagnij_nazwe_zmiennej(formula, nazwa = as.character(substitute(formula)),
  czyJednaZmienna = TRUE)
}
\arguments{
\item{formula}{parametr klasy 'formula'.}

\item{nazwa}{ciąg znaków wykorzystywany w komunikcie błedu.}

\item{czyJednaZmienna}{zmienna logiczna. Jeżeli wartość wynosi TRUE to funkcja zwraca błąd dla formuł prawostronnych, które
zawierają więcej niż jedną zmienną.}
}
\value{
wektor ciągów znakowych, które są nazwami prawostronnych zmiennych formuły.
#
}
\description{
Funkcja pomocnicza, która zwraca nazwy prawostronnych zmiennych formuły. Jeżeli formuła jest obustronna to funkcja zwraca błąd.
}

