% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obliczanie_ewd.R
\name{zgadnij_zm_egz_we}
\alias{zgadnij_zm_egz_we}
\title{Obliczanie EWD}
\usage{
zgadnij_zm_egz_we(model)
}
\arguments{
\item{model}{model regresji, typowo klasy \code{lm} lu \code{lmer}}
}
\value{
ciąg znaków (nazwa zmiennej)
}
\description{
Funkcja na podstawie modelu stara się zgadnąć nazwę zmiennej opisującej
wyniki egzaminu "na wejściu".
}
\details{
Heurystyka szukania jest następująca: 1) weź zmienne występujące w formule
modelu, 2) usuń z nich: pierwszą (tj. zależną) oraz te, które wedle
wszelkiego prawdopodobieństwa opisują płeć, dysleksję, rok zdawania egzaminu
czy długość toku kształcenia, 3) spośród pozostałych wybierz tą, która
najczęściej występuje w nazwach kolumn model.matrix.
}
