% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dane_do_kalkulatora}
\alias{dane_do_kalkulatora}
\title{Kalkulator EWD}
\usage{
dane_do_kalkulatora(modele, mapowanie, dane)
}
\arguments{
\item{modele}{lista modeli regresji MNK}

\item{mapowanie}{lista zawierająca mapowania wyników surowych egzaminów na skalę 100;15}

\item{dane}{data frame z danymi używanymi do estymacji ww. modeli}
}
\value{
lista
}
\description{
Funkcja zwraca listę zawierającą zestawienia, które trzeba wkleić do odpowiedniego
arkusza kalkulacyjnego i przesłać do Blue Bricka, aby mogli zaimplementować modele
w Kalkulatorze.
}
\details{
Dane muszą być przekazane oddzielnym argumentem, aby uprościć sobie sprawę
z wykrywaniem, jakie były najniższe wartości egzaminu na wejściu występujące w danych
(jeśli w formule modelu użyto funkcji \code{poly()}, jest to niebanalne).
}

