% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnostyka_pelna.R
\name{diagnostyka_pelna}
\alias{diagnostyka_pelna}
\title{Pełna diagnostyka wielomianow}
\usage{
diagnostyka_pelna(dane, stopnie, typModelu, zmEgzWej, zmEgzWyj,
  zmGrupujace = NULL, efektLosowy = NULL, zmDyslWej = NULL,
  zmDyslWyj = NULL, katalogRysunki = NULL, porownModeliPlik = NULL,
  tytulDiagnostyka = "", maska = NULL)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu}

\item{stopnie}{wektor liczb całkowitych określających stopnie wielomianów,
dla których należy zbudować modele.}

\item{typModelu}{zmienna określająca, jaki model ma być użyty. Dopuszczalne
wartości to "lm" oraz "lmer".}

\item{zmEgzWej}{nazwa kolumny ramki 'dane', która zawiera wyniki(umiejętności)
egzaminu wejściowego.}

\item{zmEgzWyj}{nazwa kolumny ramki 'dane', która zawiera wyniki(umiejętności)
egzaminu wyjściowego.}

\item{zmGrupujace}{wektor ciągów znakowych określających zmienne modelu, na
podstawie których określone są grupy, w których wykonywane są obliczenia oraz
dla których są sporządzane wykresy}

\item{efektLosowy}{nazwa kolumny ramki 'dane', która określa efekt losowy.
Parametr ignorowany dla modelu 'lm'.}

\item{zmDyslWej}{nazwa kolumny ramki 'dane', która zawiera informacje
o dysleksji zgłoszonej dla egzaminu wejściowego.}

\item{zmDyslWyj}{nazwa kolumny ramki 'dane', która zawiera informacje
o dysleksji zgłoszonej dla egzaminu wyjściowego.}

\item{katalogRysunki}{ścieżka do katalogu do zapisu wykresów diagnostycznych.}

\item{porownModeliPlik}{ścieżka do pliku do zapisu porównania modeli.
Parametr użyty tylko, gdy rozważamy wiele modeli.}

\item{tytulDiagnostyka}{tytuł wykresu diagnostycznego rysowanego, gdy
rozważany jest tylko jeden model.}

\item{maska}{wektor logiczny określający, które obserwacje z ramki dane
powinny zostać użyte do wyestymowania modeli.}
}
\value{
funkcja zwraca listę modeli lub model (w przypadku podania parametru
'stopnie' o długości 1).
}
\description{
Funkcja wykonuje pełną diagnostykę modeli EWD.
}

