% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnostyka_wielomianow.R
\name{diagnostyka_wielomianow}
\alias{diagnostyka_wielomianow}
\title{Diagnostyka wielomianow}
\usage{
diagnostyka_wielomianow(model, zmWielomian, zmGrupujace = NULL,
  folderWykresy = NULL, wykresyFun = png, wykresyParam = NULL,
  przew_nparPar = NULL, smooothScatterPar = NULL)
}
\arguments{
\item{model}{obiekt klasy 'lm' lub 'lmerMod'}

\item{zmWielomian}{ciąg znaków określający zmienną wielomianu}

\item{zmGrupujace}{wektor ciągów znakowych określających zmienne modelu, na podstawie których określone są grupy, w których wykonywane są obliczenia oraz dla których są sporządzane wykresy}

\item{folderWykresy}{ścieżka do folderu, gdzie powinny być zapisane wykresy. Jeżeli jest równa NULL to wykresy nie są zapisywane.}

\item{wykresyFun}{funkcja rysująca wykresy. Domyślnie jest to \code{png}.}

\item{wykresyParam}{lista parametrów funkcji rysującej wykresy.}

\item{przew_nparPar}{lista z parametrami funkcji \code{\link{przew_npar}}.}

\item{smooothScatterPar}{lista z parametrami funkcji \code{\link[graphics]{smoothScatter}}.}
}
\value{
funkcja zwraca listę z obliczeniami dla modelu.
}
\description{
Funkcja oblicza link test, BIC oraz współczynnik R2 (w przypadku modeli 'lm') lub dekompozycje wariancji (dla modeli 'lmerMod'),
sprawdza czy składowa wielomianowa modelu jest rosnąca oraz rysuje wykresy reszt modelu z wykresem przewidywań wielomianem
i regesją nieparametryczną.
}

