% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obliczanie_ewd.R
\name{ewd_es}
\alias{ewd_es}
\title{Obliczanie EWD (Kalkulator)}
\usage{
ewd_es(model, idSzkoly, noweDane = NULL)
}
\arguments{
\item{model}{model klasy \code{lm}}

\item{idSzkoly}{data frame zawierający kolumnę z identyfikatorami szkół}

\item{noweDane}{opcjonalnie data frame zawierający nowe dane (tj. nie te,
na podstawie których został wyestymowany model), dla których mają zostać
obliczone przewidywania, reszty i w efekcie wartości wskaźników}
}
\value{
data frame zawierający oszacowania dla poszczególnych szkół: EWD,
błędu standardowego EWD, średniego wyniku końcowego i jego błędu standardowego
}
\description{
Funkcja zwraca oszacowania EWD i ich błędy standardowe obliczane jako średnia
z reszt regresji MNK oraz średnie wyniki egzaminu na wyjściu i ich błędy
standardowe obliczane jako błąd standardowy średniej z prostej próby losowej.
}
\details{
Ponieważ obiekt z modelem regresji MNK nie zawiera informacji o przydziale
uczniów do szkół, musi ona zostać podana oddzielnym parametrem
(\code{id_szkoly}). Musi to być data frame, zawierający kolumnę z id szkoły
(i najlepiej tylko nią jedną), utworzony przez usunięcie kolumn z tego samego
obiektu, z którego był estymowany model \code{x}. Łączenie jest
przeprowadzane po nazwach wierszy (\code{model.frame(x)})' i (\code{id_szkoly}).
}
