% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obliczanie_ewd.R
\name{ewd_me_ssr}
\alias{ewd_me_ssr}
\title{Obliczanie EWD}
\usage{
ewd_me_ssr(x, noweDane = NULL)
}
\arguments{
\item{x}{model klasy \code{lmerMod}}

\item{noweDane}{ramka danych z danymi uczniów dla szkół. Jeżeli NULL to
funkcja liczy EWD na danych pobranych z modelu.}
}
\value{
data frame z potencjalnym atrybutem \code{noweDane}, który zawiera
parametr \code{noweDane}
}
\description{
Funkcja zwraca oszacowania EWD i ich "błędy standardowe" wykorzystując metodę
"ściągniętych średnich reszt".
}
\details{
Funkcja powstała jako awaryjne rozwiązanie w sytuacji, gdy
\code{ranef(., condVar=TRUE)} nie dawało dobrych oszacowań "błędów
standardowych" ze względu na błąd autorów pakietu \code{lme4}.
}
