% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_ewd.R
\name{konwertujNalmeEWD}
\alias{konwertujNalmeEWD}
\title{Konwersja obiektu lme4 na lmeEWD}
\usage{
konwertujNalmeEWD(objectLME4)
}
\arguments{
\item{objectLME4}{model klasy \code{lmerMod}}
}
\value{
obiekt klasy \code{lmeEWD}
}
\description{
Funkcja tworzy obiekt klasy \code{lmeEWD} na podstawie obiektu \code{lmerMod}.
}
