% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_ewd.R
\name{predict.lmeEWD}
\alias{predict.lmeEWD}
\title{Przewidywanie dla obiektu 'lmeEWD'}
\usage{
\method{predict}{lmeEWD}(object, ...)
}
\arguments{
\item{object}{model klasy \code{lmeEWD}}

\item{...}{parametry: \code{newdata} - ramka danych, dla których należy
obliczyć przewidywania oraz \code{zLosowymi} -  wartość logiczna określająca,
czy uwzględniać w przewidywaniach efekt losowy}
}
\value{
wektor liczb
}
\description{
Funkcja tworzy obiekt klasy 'lmeEWD' na podstawie obiektu 'lmerMod'.
}
