% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przeksztalc_skale_zmiennych.R
\name{przeksztalc_skale_zmiennych}
\alias{przeksztalc_skale_zmiennych}
\title{Wyliczanie EWD}
\usage{
przeksztalc_skale_zmiennych(dane, czesciEgzWy, sufiksWyniki)
}
\arguments{
\item{dane}{data frame z danymi, typowo wczytany do R wynik działania
funkcji \code{\link[EWDdane]{przygotuj_dane_do_ewd}}}

\item{czesciEgzWy}{wektor ciągów znaków - kody konstruktów, które będą
w modelach EWD zmiennymi zależnymi}

\item{sufiksWyniki}{ciąg znaków: "irt", "norm" lub "suma"; określa, które
wyniki będą używane do obliczania wskaźników i mają zostać przekształcone}
}
\value{
data frame
}
\description{
Funkcja przekształca skale zmiennych z surowymi lub wyskalowanymi wynikami
egzaminów ze 100;15 na 0;1, lub odwrotnie, adkewatnie do tego, jak typowo
są one wykorzystywane w modelach EWD.
}

