% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyliczanie_ewd.R
\name{przygotuj_wsk_ewd}
\alias{przygotuj_wsk_ewd}
\title{Wyliczanie EWD}
\usage{
przygotuj_wsk_ewd(modele, dane, danePominiete = NULL, skale = NULL,
  powiazaniaPrzedmiotow = NULL)
}
\arguments{
\item{modele}{lista modeli klasy \code{lm} lub \code{lmerMod}}

\item{dane}{data frame z dodatkowymi danymi, w szczególności zawierający
kolumnę \code{lu_wszyscy}, przypisanymi tym samym obserwacjom, na podstawie
których estymowane były modele podane parametrem \code{modele}}

\item{danePominiete}{opcjonalnie data frame o takiej samej strukturze, jak
przekazany argumentem \code{dane}, zawierający dane szkół, które nie zostały
wykorzystane na etapie estymacji modelu, ale chcemy teraz wyliczyć dla nich
wartości wskaźników}

\item{skale}{data frame zawierający informacje o skalowaniach (i skalach),
z których pochodzą zmienne z oszacowaniami umiejętności, zawarte w danych;
typowo atrybut \code{skale} obiektu klasy \code{daneWyskalowane}
(zwracanego przez funkcję \code{\link[EWDdane]{przygotuj_dane_do_ewd}})}
}
\value{
lista data frame'ów
}
\description{
Funkcja na podstawie listy modeli, data frame'a z danymi, na podstawie
których zostały one wyliczone oraz ew. data frame'a o takiej samej
strukturze zawierającego dane nie użyte wcześniej do estymacji modeli,
przygotowuje zestawienia z wartościami wskaźników EWD.
}

