% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pseudoR2}
\alias{pseudoR2}
\title{Wyliczenia do wspołczynnika R-kwadrat.}
\usage{
pseudoR2(model)
}
\arguments{
\item{model}{parametr klasy 'lmerMod' lub 'lm'.}
}
\value{
Funkcja zwraca liczbę lub macierz w zależności od klasy obiektu model.
}
\description{
Funkcja dla modeli 'lm' zwraca współczynnik R-kwadrat, dla modeli 'lmerMod' zwraca tabicę dekompozycji wariancji na efekty stałe i losowe.
}

