% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wyliczanieEWD.R
\name{sr_wy}
\alias{sr_wy}
\title{Wyliczanie EWD}
\usage{
sr_wy(model, ewd)
}
\arguments{
\item{model}{model klasy \code{lmerMod}}

\item{ewd}{data frame będący wynikiem działania funkcji \code{\link{ewd_me}} lub \code{\link{ewd_me_ssr}}.
Jeżeli ten parametr zawiera atrybut noweDane to funkcja wykonuje wyliczenia na tym atrybucie.}
}
\value{
data frame
}
\description{
Funkcja zwraca oszacowania średnich wyników na wyjściu i ich "błędy standardowe"
oraz korelacje z EWD wykorzystując metodę "z jednego modelu".
}

