% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelki.r
\name{tabelka_ld}
\alias{tabelka_ld}
\title{Liczba zdających egzaminy.}
\usage{
tabelka_ld(x, nazwyZmWynikiEgzWy, kodyCzesciEgzWe, nazwaEgzWy = "egz. wy.",
  nazwaEgzWe = "egz. we.")
}
\arguments{
\item{x}{data frame z danymi potrzebnymi do przygotowania zestawienia}

\item{nazwyZmWynikiEgzWy}{wektor ciągów znaków zawierający nazwy zmiennych
z wynikami poszczególnych części egzaminu "na wyjściu", dla których ma być
przygotowane zestawienie}

\item{kodyCzesciEgzWe}{wektor ciągów znaków - kody części egzaminu
"na wejściu" (np. "gm_m", "s")}

\item{nazwaEgzWy}{ciąg znaków - nazwa egzaminu na wyjściu (do nagłówka tabeli)}

\item{nazwaEgzWe}{ciąg znaków - nazwa egzaminu na wejściu (do nagłówka tabeli)}
}
\value{
data frame
}
\description{
Funkcja zwraca tabelę z zestawieniem liczby podchodzących do dwóch egzaminów
(na wejściu i na wyjściu), z wyszczególnieniem liczby laureatów i dyslektyków.
}
